Revert "Merge pull request #35 from santino2003/debBranch"
This reverts commit 38c320fa9411f166f40044f1bd259d824d953698, reversing
changes made to b279e0d7e86bc82658f0d3949d52e1803ae26b75.

+++ src/assets/js/ajusteInventario.js
@@ -93,7 +93,7 @@ $(document).ready(function() {
         }
 
         // Validar que la cantidad sea mayor a 0
-        const cantidad = parseFloat($('#cantidad').val());
+        const cantidad = parseInt($('#cantidad').val());
         if (!cantidad || cantidad <= 0) {
             Swal.fire({
                 icon: 'warning',
